call the lalr parser's error procedure with a useful argument
Patch by: Ludovic Courtès <ludo@gnu.org>

The error procedure that is passed to parsers generated by `lalr-parser'
has limited utility since, upon parse error, it is passed an error
message and the *name* of the offending token (i.e., a symbol), not the
token itself.  The attached patch changes the latter so that the
complete token (i.e., a pair) is passed, not just its name (which is the
`car' of said pair).

To illustrate the problem, consider this example.  Suppose you're
getting a parse error due to, say, an `identifier' token occurring in
the wrong place.  In this case, the error procedure is currently invoked
like this:

  (errorp "PARSE ERROR : unexpected token : " 'identifier)

Thus you get very little information as to what led to the parse error:
you don't know what the offending identifier is, where it's located in
the source file, etc.

With the patch, the full token (which could include location
information) is passed, leading to, e.g.:

  (errorp "PARSE ERROR : unexpected token : "
          '(identifier "foo" ((line . 2) (column . 3) (file . "foo.x"))))

This is an incompatible change, but probably not a disturbing one since
error procedures could currently not do much beyond the standard
`error'.